+++ src/@components/JoinPage/EmailAuthenticationPage/style.ts
