+++ src/@components/index.ts
@@ -2,9 +2,9 @@ export { default as BookmarkPage } from "./BookmarkPage";
 export { default as CardCollectionPage } from "./CardCollectionPage";
 export { default as CategoryPage } from "./CategoryPage";
 export { default as Error404Page } from "./Error404Page";
-export { default as EmailAuthenticationPage } from "./JoinPage/ EmailAuthentication";
-export { default as EmailConfirmPage } from "./JoinPage/EmailConfirm";
-export { default as JoinAgreePage } from "./JoinPage/JoinAgree";
+export { default as AgreePage } from "./JoinPage/AgreePage";
+export { default as EmailAuthenticationPage } from "./JoinPage/EmailAuthenticationPage";
+export { default as EmailConfirmPage } from "./JoinPage/EmailConfirmPage";
 export { default as LoginPage } from "./LoginPage";
 export { default as MainPage } from "./MainPage";
 export { default as MyPage } from "./MyPage";
